Fixed: Issue with Connecting CSS Root Color in Style.css
Problem Description:
There was an issue in the style.css file where the background font was not changing to the primary color as defined in the root.

Cause:
The issue was caused by an extra semicolon in the CSS root color definition.

Solution:
The problem was resolved by deleting the extra semicolon from the CSS file.

Changes Made:
Deleted the extra semicolon from the CSS root color definition.

Result:
The background font now correctly changes to the primary color as defined in the root.

On branch main
Your branch is up to date with 'origin/main'.

Changes to be committed:
	modified:   assets/css/style.css