bug: Module specifier bug in JavaScript
Problem Description:
A bug was identified in the JavaScript file. The console error message was 'Uncaught TypeError: Failed to resolve module specifier "assets/js/confetti.js". Relative references must start with either "/", "./", or "../"'. This issue was causing the JavaScript to fail in its entirety.

Steps to reproduce:
1. Setup the game.
2. Select player and computer.
3. Try to reset all functions and trigger the confetti functionality.

Expected result:
The confetti.js file should import successfully and the JavaScript functions should work as expected.

Actual result:
The console throws an error about failing to resolve the module specifier and the JavaScript doesn't work.

Proposed solution:
The path to the confetti.js file should be reviewed and corrected as necessary. If the file is located in the same directory as the script that's trying to import it, the path should start with './'. If it's located in a parent directory, the path should start with '../'.

On branch main
Your branch is up to date with 'origin/main'.

Changes to be committed:
	modified:   assets/css/style.css
	new file:   assets/js/confetti.js
	modified:   assets/js/script.js
	modified:   index.html