Bug: Aligning Reset Icon with Rules Button
Issue Description:
I have stumbled upon a layout issue concerning the reset icon and the rules button. These two elements, which are supposed to be aligned in the same row, are not doing so correctly. The rules button, in particular, is appearing beneath the reset icon instead of being aligned in the same row.

Attempted Solution and Persisting Problem:
An initial solution was attempted to address this issue, where both elements were placed within the same container (a div with the id 'footer-panel') and flexbox principles were applied for alignment. Techniques such as 'flex-direction' and 'flex-wrap' were utilized, as part of the flexbox properties. However, despite these efforts, the alignment problem remains.

Resource: https://css-tricks.com/snippets/css/a-guide-to-flexbox/

On branch main
Your branch is up to date with 'origin/main'.

Changes to be committed:
	modified:   assets/css/style.css
	new file:   assets/images/rule_diagram.png
	modified:   index.html