Fix Text Alignment for Rules Text
Problem:
The issue was identified in the 'rules' modal that appears upon clicking the 'rules' button. The 'confetti' text sentences were misaligned with the rest of the paragraph.

Solution:
The issue was successfully resolved by adding the following code snippet to CSS and removing the <br/> tags from index.html:

```
.rules-text li {
max-width: 30em;
text-align: justify;
}

```

This aligns the text correctly within the paragraph across all screen sizes, providing a consistent user experience.

On branch main
Your branch is up to date with 'origin/main'.

Changes to be committed:
	modified:   assets/css/style.css
	modified:   assets/js/script.js
	modified:   index.html